// src/types/TypeReport.ts
export enum TypeReport {
    BEST_LAP = 'BEST_LAP',
    COUNT_LAPS = 'COUNT_LAPS',
    BEST_PIT_STOP = 'BEST_PIT_STOP',
    POSITION_SPORTSMEN = 'POSITION_SPORTSMEN',
    ROUND_GROUPS = 'ROUND_GROUPS',
    ROUND_GROUPS_LAPS = 'ROUND_GROUPS_LAPS',
    MX_RESULTS = 'MX_RESULTS',
    MX_LAPS = 'MX_LAPS'
}